Large bug fix
Due to the grammar passed into instaparse, input with ten or more spaces
roughly would cause out of memory errors as the parser attempted to
recursively store a large number of possible parses. It seems like this
could take up an infinite amount of memory and so we now simply replace
all spaces with a single space. In addition, as a safeguard, we now
prevent the parsing function from taking longer than one second to run
and log out any strings that can't be parsed.